MYG000-98 - Updated Add Manual Vehicle Method

+++ BK/src/main/java/com/searshc/mygarage/entities/FamilyVehicle.java
@@ -48,7 +48,7 @@ public class FamilyVehicle extends AbstractEntity implements Serializable {
     private String color;
 
     @Column(name = "mileage", nullable = false)
-    private int mileage;
+    private double mileage;
 
     /**
      * The FamilyVehicle name
@@ -75,7 +75,7 @@ public class FamilyVehicle extends AbstractEntity implements Serializable {
      * @param mileage: Optional
      * @param name: Optional
      */
-    public FamilyVehicle(final Vehicle vehicle, final Long familyId, final Long tangibleId, final String color, final int mileage, final String name) {
+    public FamilyVehicle(final Vehicle vehicle, final Long familyId, final Long tangibleId, final String color, final double mileage, final String name) {
         this.vehicle = Validate.notNull(vehicle, "The Vehicle cannot be null");
         this.familyId = familyId;
         this.tangibleId = tangibleId;
@@ -144,16 +144,16 @@ public class FamilyVehicle extends AbstractEntity implements Serializable {
     /**
      * @return the mileage
      */
-    public int getMileage() {
+    public double getMileage() {
         return mileage;
     }
 
     /**
      * @param mileage the mileage to set
      */
-    public void setMileage(int mileage) {
-        Validate.isTrue(mileage >= 0, "The Mileage cannot be lower than 0");
-        this.mileage = mileage;
+    public void setMileage(double mileage) {
+    	Validate.isTrue(mileage >= 0, "The Mileage cannot be lower than 0");
+    	this.mileage = mileage;
     }
 
     /**
